requirements.txt
Flask
pillow
easyocr
gspread
pytesseract
werkzeug